requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
defusedxml==0.6.0
distlib==0.3.1
Django==2.0
djangorestframework==3.11.0
docopt==0.6.2
filelock==3.0.12
gunicorn==19.7.1
idna==2.10
importlib-metadata==3.4.0
importlib-resources==5.0.0
isort==5.6.4
joblib==0.17.0
Js2Py==0.70
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandas==0.25.3
pdfkit==0.6.1
Pillow==5.0.0
protobuf==3.15.5
psycopg2==2.7.3.2
PyAudio==0.2.11
pyjsparser==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python-http-client==3.2.7
python3-openid==3.2.0
pyttsx3==2.90
pytube3==9.6.4
pytz==2020.5
pywin32==300
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
scipy==1.5.4
sendgrid==6.4.5
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
stripe==2.55.1
toml==0.10.2
tqdm==4.54.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
views==0.3
virtualenv==15.1.0
whitenoise==3.3.1
wkhtmltopdf==0.2
wrapt==1.12.1
youtube-dl==2021.5.16
youtube-python==1.0.13
zipp==3.4.0
